Updated the changes for proxy
