+++ src/main/java/uk/gov/hmcts/reform/em/test/idam/DeleteUserApi.java
@@ -6,12 +6,11 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import uk.gov.hmcts.reform.em.FooConfiguration;
 
 @FeignClient(
     name = "idam-test-api-delete",
     url = "${idam.api.url}/testing-support",
-        configuration = FooConfiguration.class
+        configuration = IdamConfiguration.class
 )
 @ConditionalOnProperty("idam.api.url")
 public interface DeleteUserApi {
